Refactored data models and removed DataRecords parsing from Recipes
