+++ DruidsCornerApp/Models/DruidsCornerApi/IndexedDb/IndexedHopDb.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.IndexedDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.IndexedDb
 {
     /// <summary>
     /// Encodes basic information about BrewDog's beer recipe.
