+++ DruidsCornerApp/Models/DruidsCornerApi/IndexedDb/IndexedMaltDb.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace DruidsCornerAPI.Models.DiyDog.IndexedDb
+namespace DruidsCornerApp.Models.DruidsCornerApi.IndexedDb
 {
     /// <summary>
     /// Encodes basic information about BrewDog's beer recipe.
