+++ DruidsCornerApp/Models/DruidsCornerApi/IndexedDb/ReversedPropMapping.cs
@@ -1,5 +1,5 @@
 
-namespace DruidsCornerAPI.Models.DiyDog.IndexedDb
+namespace DruidsCornerApp.Models.DruidsCornerApi.IndexedDb
 {
 
     /// <summary>
