+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/AllRecipesCollection.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Collection of Recipes.
