+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/Hop.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Hop ingredient
