+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/Ingredients.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Ingredients datastructure
