+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/MethodTimings.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Method Timings characteristics
