+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/Recipe.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Available packaging types for the beers
@@ -88,17 +88,6 @@
         /// Most probable packaging type of this beer. Guessed from the overall shape of the beer's picture.
         /// </summary>
         public PackagingType PackagingType { get; set; } = PackagingType.Bottle;
-
-        /// <summary>
-        // /// Beer's image (can either point to a local file source or a cloud record.)
-        // /// </summary>
-        // public DataRecord Image { get; set; } = new FileRecord();
-        //
-        // /// <summary>
-        // /// Beer's original PDF page. Can either point to a local file source or a cloud record.
-        // /// This PDF page might be requested for the end user to countercheck some values, parsing is not 100% reliable! 
-        // /// </summary>
-        // public DataRecord PdfPage { get; set; } = new FileRecord();
 
         /// <summary>
         /// Optional list of food pairings
