+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/Temperature.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Temperature instruction for Mash
