+++ DruidsCornerApp/Models/DruidsCornerApi/RecipeDb/Twist.cs
@@ -1,4 +1,4 @@
-﻿namespace DruidsCornerAPI.Models.DiyDog.RecipeDb
+﻿namespace DruidsCornerApp.Models.DruidsCornerApi.RecipeDb
 {
     /// <summary>
     /// Additional so-called "twist" ingredients.
